+++ packages/media-parser/src/test/stsd.test.ts
@@ -52,7 +52,6 @@ test('Should be able to parse a STSD audio box correctly', async () => {
 			fields: {},
 			nullifySamples: false,
 			supportsContentRange: true,
-			nextTrackIndex: 0,
 		}),
 		signal: null,
 		fields: {},
@@ -225,7 +224,6 @@ test('Should be able to parse a STSD video box correctly', async () => {
 
 			nullifySamples: false,
 			supportsContentRange: true,
-			nextTrackIndex: 0,
 		}),
 
 		signal: null,
